Refactored userController tests and mocked setup
